refactor(questionnaires): extract QuestionnaireFillForm from event fill page
Move the questionnaire form body (answer state, conditional-question
wiring, validation, payload building and the four render snippets) out of
the event questionnaire fill page into a reusable component so other
routes can render the same form.

No behavior change: identical markup, validation flow and submission
payload. The page keeps its API calls, header copy, auto-accepted card
and navigation, and now feeds the form via props.
